+++ include/BDSParticleDefinition.hh
@@ -32,37 +32,51 @@ class G4ParticleDefinition;
  * Used to calculate momentum and kinetic energy and associate any other information
  * required along with a G4ParticleDefinition instance.
  *
+ * Bad construction can throw a BDSException.
+ *
  * @author Laurie Nevay
  */
 
 class BDSParticleDefinition
 {
 public:
-  /// Constructor requires G4 particle definition as well as user specified
-  /// total energy. From this the momentum and kinetic energy are calculated
-  /// and checked to be valid - ie sufficient total energy. The optional ion
-  /// definition may be supplied to override the charge of the ion. ffact is
+  /// Constructor requires G4 particle definition as well as one of total energy,
+  /// kineticEnergy or momentum. Only 1 of these can be non-zero (will throw
+  /// exception otherwise). The others are calculated using the mass. The optional
+  /// ion definition may be supplied to override the charge of the ion. ffact is
   /// typically 1 or -1 to flip the rigidity for convention.
   BDSParticleDefinition(G4ParticleDefinition* particleIn,
 			G4double              totalEnergyIn,
+			G4double              kineticEnergyIn,
+			G4double              momentumIn,
 			G4double              ffact,
 			BDSIonDefinition*     ionDefinitionIn = nullptr);
-
+  
   /// Alternative constructor for when we don't have access to the particle table
   /// information. G4ParticleDefinition can be updated later. Developer
   /// responsibility to ensure this matches the contents of this class.
-  /// ffact is typically 1 or -1 to flip the rigidity for convention.
+  /// ffact is typically 1 or -1 to flip the rigidity for convention. As before
+  /// only one of totalEnergy, kineticEnergy or momentum should be specified.
   BDSParticleDefinition(G4String          nameIn,
 			G4double          massIn,
 			G4double          chargeIn,
 			G4double          totalEnergyIn,
+			G4double          kineticEnergyIn,
+			G4double          momentumIn,
 			G4double          ffact,
-		       	BDSIonDefinition* ionDefinitionIn = nullptr);
+			BDSIonDefinition* ionDefinitionIn = nullptr);
 
   /// Copy constructor specified as we have to copy the ionDefinition if it exists.
   BDSParticleDefinition(const BDSParticleDefinition& other);
   
   ~BDSParticleDefinition();
+
+  /// Check in order whether totalEnergy, kineticEnergy or momentum are specified and
+  /// copy that parameter to the member varible. Then calculate the other two and set
+  /// them to the member variables.
+  void SetEnergies(G4double totalEnergyIn,
+		   G4double kineticEnergyIn,
+		   G4double momentumIn);
 
   /// Update the G4 particle definition member. Developer responsibility to ensure
   /// this matches the contents of the class.
@@ -80,6 +94,7 @@ public:
   inline G4double Gamma()         const {return gamma;}
   inline G4double Beta()          const {return beta;}
   inline G4double BRho()          const {return brho;}
+  inline G4double FFact()         const {return ffact;}
   inline G4bool   IsAnIon()       const {return ionDefinition != nullptr;}
   inline G4bool   NElectrons()    const {return ionDefinition != nullptr ? ionDefinition->NElectrons() : 0;}
   /// @}
@@ -95,7 +110,7 @@ private:
   void CalculateMomentum();
 
   /// Calculate and set rigidity based on charge and momentum.
-  void CalculateRigidity(const G4double& ffact);
+  void CalculateRigidity(const G4double& ffactIn);
 
   /// Calculate and set lorentz factors.
   void CalculateLorentzFactors();
@@ -112,6 +127,7 @@ private:
   G4double gamma;          ///< Relativistic gamma.
   G4double beta;           ///< Relativistic beta.
   G4double brho;           ///< Particle rigidity.
+  G4double ffact;          ///< Flip factor.
 };
 
 #endif